Moved algorithm.ts interface into better position (folder layout)
